+++ qml/components/HeroHeader.qml
@@ -18,6 +18,7 @@ Item {
     required property bool configOpen
     required property bool hintMode
     required property color accent
+    required property color tipBackground
     required property double dayTotal
     required property string activeDayKey
     required property string activeDayLabel
@@ -175,6 +176,15 @@ Item {
                 gearSpin.restart();
             heroHeader.configToggled();
         }
+    }
+
+    ScreenTip {
+        foreground: heroHeader.foreground
+        fontFamily: heroHeader.fontFamily
+        tipBackground: heroHeader.tipBackground
+
+        hovered: configGearMouse.containsMouse
+        tipText: "Settings"
     }
 
     HintBadge {